Added a BitMap Contract which should be fullfilled by implementations, renamed the SPLArray based Bitmap implementation into SPLArrayBitMap. For compatibility both BitMap/BitSet won't be removed for now
